Website about me with Spring
Tthis project shows the possibility of using Spring, H2 and Hibernate with Java. The website was created using Maven in the InteliJ environment.
The frontend was written by me to better understand the entire operation of the page creation mechanism. Inside You can find my Cv and download it,
there is also a form using spring and javascript that writes to the local database. I encourage you to watch this project in particular.